+++ metadata-ingestion/setup.py
@@ -202,12 +202,7 @@ setuptools.setup(
             "datahub-kafka = datahub.ingestion.sink.datahub_kafka:DatahubKafkaSink",
             "datahub-rest = datahub.ingestion.sink.datahub_rest:DatahubRestSink",
         ],
-        "apache_airflow_provider": [
-            "provider_info=datahub.integrations.airflow.get_provider_info:get_provider_info"
-        ],
-        "airflow.plugins": [
-            "datahub = datahub.integrations.airflow.get_provider_info:DatahubAirflowPlugin"
-        ],
+        "apache_airflow_provider": ["provider_info = datahub_provider:get_provider_info"],
     },
     # Dependencies.
     install_requires=list(base_requirements | framework_common),
