+++ metadata-ingestion/src/datahub_provider/hooks/datahub.py
@@ -1,0 +1,5 @@
+from datahub.integrations.airflow.hooks import (  # noqa: F401
+    DatahubGenericHook,
+    DatahubKafkaHook,
+    DatahubRestHook,
+)
