+++ metadata-ingestion/src/datahub_provider/operators/datahub.py
@@ -1,0 +1,4 @@
+from datahub.integrations.airflow.operators import (  # noqa: F401
+    DatahubBaseOperator,
+    DatahubEmitterOperator,
+)
