+++ metadata-ingestion/tests/unit/test_airflow.py
@@ -19,9 +19,9 @@ except ModuleNotFoundError:
 
 import datahub.emitter.mce_builder as builder
 from datahub.integrations.airflow.entities import Dataset
-from datahub.integrations.airflow.get_provider_info import get_provider_info
 from datahub.integrations.airflow.hooks import DatahubKafkaHook, DatahubRestHook
 from datahub.integrations.airflow.operators import DatahubEmitterOperator
+from datahub_provider import get_provider_info
 
 lineage_mce = builder.make_lineage_mce(
     [
